+++ contract/eoscrazytown/eoscrazytown.abi
@@ -1,5 +1,5 @@
 {
-  "____comment": "This file was generated by eosio-abigen. DO NOT EDIT - 2018-11-06T05:48:05",
+  "____comment": "This file was generated by eosio-abigen. DO NOT EDIT - 2018-11-06T09:42:52",
   "version": "eosio::abi/1.0",
   "types": [{
       "new_type_name": "time",
@@ -62,7 +62,7 @@
           "type": "uint64"
         },{
           "name": "earnings_per_share",
-          "type": "uint64"
+          "type": "uint128"
         }
       ]
     },{
